refactor item-classes
